+++ src/joynet/serialize.h
@@ -1,6 +1,7 @@
 #ifndef JOYNET_SERIALIZE_H
 #define JOYNET_SERIALIZE_H
 
+#include <stdint.h>
 #include "joynet.h"
 
 JOYNET_SERIAL_DATA * joynet_create_serial_data(void);
@@ -10,10 +11,10 @@ void joynet_read(JOYNET_SERIAL_DATA * dp, char * dest, int num);
 void joynet_write(JOYNET_SERIAL_DATA * dp, char * dest, int num);
 void joynet_getc(JOYNET_SERIAL_DATA * dp, char * c);
 void joynet_putc(JOYNET_SERIAL_DATA * dp, char c);
-void joynet_getw(JOYNET_SERIAL_DATA * dp, short * c);
-void joynet_putw(JOYNET_SERIAL_DATA * dp, short c);
-void joynet_getl(JOYNET_SERIAL_DATA * dp, long * c);
-void joynet_putl(JOYNET_SERIAL_DATA * dp, long c);
+void joynet_getw(JOYNET_SERIAL_DATA * dp, int16_t * c);
+void joynet_putw(JOYNET_SERIAL_DATA * dp, int16_t c);
+void joynet_getl(JOYNET_SERIAL_DATA * dp, int32_t * c);
+void joynet_putl(JOYNET_SERIAL_DATA * dp, int32_t c);
 int joynet_get_serial_size(JOYNET_SERIAL_DATA * dp);
 
 #endif
